Added memory mode for AimBot for better accuracy
* Mouse mode still available in config as an option
* Scaled AimSpeed for consistent settings between the two modes
* Added tick rate counter to avoid overadjusting and causing untrusted

+++ src/utils/hmath.h
@@ -66,7 +66,7 @@ inline float HMath::GetFov(const Vector& eyePos, const Vector& localHeadPos, con
 inline Vector HMath::VectorAngles(const Vector& dir)
 {
     Vector angles = dir;
-    float hyp = std::sqrt((dir.x * dir.x) + (dir.y * dir.y));
+    double hyp = std::sqrt((dir.x * dir.x) + (dir.y * dir.y));
     angles.x = (float) std::atan(dir.z / hyp) * HMath::RADPI;
     angles.y = (float) std::atan(dir.y / dir.x) * HMath::RADPI;
     if (dir.x >= 0.0) {
